pyproject.toml
[tool.poetry]
name = "pythoncesar"
version = "0.1.0"
description = ""
authors = ["CapitanCesar"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
matplotlib = "^3.10.0"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
pre-commit = "^4.0.1"
ruff = "^0.9.1"
streamlit = "^1.41.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
